+++ CHANGELOG.md
@@ -122,7 +122,33 @@ Those changes in added, changed or breaking changes, should include usage exampl
         .await
         .unwrap();
   ```
-  
+
+* Added new method `create_slashable_stake_quorum` in `avsregistry/writer` in [#340](https://github.com/Layr-Labs/eigensdk-rs/pull/340).
+
+  ```rust
+      let operator_set_param = OperatorSetParam {
+          maxOperatorCount: 10,
+          kickBIPsOfOperatorStake: 50,
+          kickBIPsOfTotalStake: 50,
+      };
+      let minimum_stake = U96::from(100);
+      let strategy_param = StrategyParams {
+          strategy: get_erc20_mock_strategy(http_endpoint.clone()).await,
+          multiplier: U96::from(1),
+      };
+      let look_ahead_period = 10;
+
+      let tx_hash = avs_writer
+          .create_slashable_stake_quorum(
+              operator_set_param,
+              minimum_stake,
+              vec![strategy_param],
+              look_ahead_period,
+          )
+          .await
+          .unwrap();
+  ```
+
 * Added new method `set_ejector` in `avsregistry/writer` in [#330](https://github.com/Layr-Labs/eigensdk-rs/pull/330).
   
   ```rust
@@ -131,6 +157,7 @@ Those changes in added, changed or breaking changes, should include usage exampl
   ````
 
 * Added new method `set_ejection_cooldown` in `avsregistry/writer` in [#337](https://github.com/Layr-Labs/eigensdk-rs/pull/337).
+
   ```rust
       let new_cooldown = U256::from(120);
       let tx_hash = avs_writer